+++ packages/augur-ui/src/modules/app/components/app.styles.less
@@ -214,6 +214,7 @@ h6 {
     @media @breakpoint-mobile {
       > div:nth-of-type(2) {
         width: 100%;
+        background: var(--color-page-background);
 
         > h1 {
           .text-24-bold;
