+++ packages/augur-ui/src/modules/common/search-sort.styles.less
@@ -22,3 +22,12 @@
   margin-top: 0.85rem;
   width: 100%;
 }
+
+:root[theme="BETTING"],
+:root[theme="SPORTS"] {
+  .SearchSort {
+    @media @breakpoint-mobile {
+      justify-content: flex-end;
+    }
+  }
+}
